Refactor. Refresh file structure
