+++ Backend/Backend.Application/AuctionData/IRepository/IAuctionRepository.cs
@@ -1,0 +1,10 @@
+﻿using Backend.Domain.Entity;
+
+namespace Backend.Application.AuctionData.IRepository;
+
+/// <summary>
+/// Интерфейс репозитория Аукциона
+/// </summary>
+public interface IAuctionRepository : IBaseRepository<Auction>
+{
+}
