--- Backend/Backend.Application/Dto/BetDto.cs
@@ -1,32 +1,0 @@
-namespace Backend.Application.Dto;
-
-/// <summary>
-/// Ставка
-/// </summary>
-public class BetDto
-{
-    /// <summary>
-    /// Уникальный идентификатор
-    /// </summary>
-    public Guid Id { get; set; }
-
-    /// <summary>
-    /// Значение ставки
-    /// </summary>
-    public decimal Value { get; set; }
-
-    /// <summary>
-    /// Уникальный идентификатор лота
-    /// </summary>
-    public Guid LotId { get; set; }
-
-    /// <summary>
-    /// Уникальный идентификатор пользователя
-    /// </summary>
-    public Guid UserId { get; set; }
-
-    /// <summary>
-    /// VВремя ставки
-    /// </summary>
-    public DateTime DateTime { get; set; }
-}
