--- Backend/Backend.Application/Interfaces/IAuctionRepository.cs
@@ -1,10 +1,0 @@
-﻿using Backend.Domain.Entity;
-
-namespace Backend.Application.Interfaces;
-
-/// <summary>
-/// Интерфейс репозитория Аукциона
-/// </summary>
-public interface IAuctionRepository : IBaseRepository<Auction>
-{
-}
