--- Backend/Backend.Application/UseCases/ChangeState/SetDateStartAuctionHandler.cs
@@ -1,36 +1,0 @@
-using Backend.Application.Interfaces;
-
-namespace Backend.Application.UseCases.ChangeState;
-
-/// <summary>
-/// Установка даты начала аукциона
-/// </summary>
-public class SetDateStartAuctionHandler
-{
-    /// <summary>
-    /// Репозиторий аукциона
-    /// </summary>
-    private readonly IAuctionRepository _auctionRepository;
-
-    /// <summary>
-    /// .ctor
-    /// </summary>
-    /// <param name="auctionRepository">Репозиторий аукциона</param>
-    public SetDateStartAuctionHandler(IAuctionRepository auctionRepository)
-    {
-        _auctionRepository = auctionRepository;
-    }
-
-    /// <summary>
-    /// Установить дату начала аукциона
-    /// </summary>
-    /// <param name="id">Уникальный идентификатор аукциона</param>
-    public async Task SetDateStartAuctionAsync(Guid id)
-    {
-        var auction = await _auctionRepository.SelectAsync(id);
-
-        auction.SetDateStart();
-
-        await _auctionRepository.UpdateAsync(auction);
-    }
-}
