--- Backend/Backend.Application/UseCases/DeleteItems/DeleteLotHandler.cs
@@ -1,32 +1,0 @@
-using Backend.Application.Interfaces;
-
-namespace Backend.Application.UseCases.DeleteItems;
-
-/// <summary>
-/// Удаление лота
-/// </summary>
-public class DeleteLotHandler
-{
-    /// <summary>
-    /// Репозиторий лота
-    /// </summary>
-    private readonly ILotRepository _lotRepository;
-
-    /// <summary>
-    /// .ctor
-    /// </summary>
-    /// <param name="lotRepository">Репозиторий лота</param>
-    public DeleteLotHandler(ILotRepository lotRepository)
-    {
-        _lotRepository = lotRepository;
-    }
-
-    /// <summary>
-    /// Удаление лота
-    /// </summary>
-    /// <param name="id">Уникальный идентификатор лота</param>
-    public async Task DeleteLotAsync(Guid id)
-    {
-        await _lotRepository.DeleteAsync(id);
-    }
-}
