--- Backend/Backend.Application/UseCases/DeleteItems/DeleteUserHandler.cs
@@ -1,32 +1,0 @@
-using Backend.Application.Interfaces;
-
-namespace Backend.Application.UseCases.DeleteItems;
-
-/// <summary>
-/// Удаление пользователя
-/// </summary>
-public class DeleteUserHandler
-{
-    /// <summary>
-    /// Репозиторий пользователя
-    /// </summary>
-    private readonly IUserRepository _userRepository;
-
-    /// <summary>
-    /// .ctor
-    /// </summary>
-    /// <param name="userRepository">Репозиторий пользователя</param>
-    public DeleteUserHandler(IUserRepository userRepository)
-    {
-        _userRepository = userRepository;
-    }
-
-    /// <summary>
-    /// Удалить пользователя
-    /// </summary>
-    /// <param name="id">Уникальный идентификатор пользователя</param>
-    public async Task DeleteUserAsync(Guid id)
-    {
-        await _userRepository.DeleteAsync(id);
-    }
-}
