--- Backend/Backend.Application/UseCases/UpdateItems/UpdateLotHandler.cs
@@ -1,55 +1,0 @@
-using Backend.Application.Dto;
-using Backend.Application.Interfaces;
-using Backend.Domain.Entity;
-
-namespace Backend.Application.UseCases.UpdateItems;
-
-/// <summary>
-/// Обновление лота
-/// </summary>
-public class UpdateLotHandler
-{
-    /// <summary>
-    /// Репозиторий аукциона
-    /// </summary>
-    private readonly IAuctionRepository _auctionRepository;
-
-    /// <summary>
-    /// .ctor
-    /// </summary>
-    /// <param name="auctionRepository">Репозиторий аукциона</param>
-    public UpdateLotHandler(IAuctionRepository auctionRepository)
-    {
-        _auctionRepository = auctionRepository;
-    }
-
-    /// <summary>
-    /// Обновить лот
-    /// </summary>
-    /// <param name="entity">Модель лота</param>
-    public async Task UpdateLotAsync(LotDto entity)
-    {
-        var auction = await _auctionRepository.SelectAsync(entity.AuctionId);
-
-        var lot = new Lot(
-            entity.Id,
-            entity.Name,
-            entity.Description,
-            entity.AuctionId,
-            entity.StartPrice,
-            entity.BuyoutPrice,
-            entity.BetStep,
-            entity.State);
-
-        var images = lot.Images.Select(image => new Image
-        {
-            Id = image.Id,
-            LotId = image.LotId,
-            Path = image.Path
-        });
-
-        auction.UpdateLot(lot.Id, lot.Name, lot.Description, lot.BetStep, images);
-
-        await _auctionRepository.UpdateAsync(auction);
-    }
-}
