+++ Backend/Backend.Application/UserData/Dto/UserDto.cs
@@ -1,0 +1,27 @@
+namespace Backend.Application.UserData.Dto;
+
+/// <summary>
+/// Пользователь
+/// </summary>
+public class UserDto
+{
+    /// <summary>
+    /// Уникальный идентификатор
+    /// </summary>
+    public Guid Id { get; init; }
+
+    /// <summary>
+    /// Имя пользователя
+    /// </summary>
+    public string Name { get; set; } = string.Empty;
+
+    /// <summary>
+    /// Почта пользователя
+    /// </summary>
+    public string Email { get; set; } = string.Empty;
+
+    /// <summary>
+    /// Пароль пользователя
+    /// </summary>
+    public string Password { get; set; } = string.Empty;
+}
